requirements.txt
Markdown==2.4
Pillow==2.3.0
Pygments==1.6
argparse==1.2.1
django-json-field==0.5.5
django-tagging==0.3.1
markup==0.2
python-dateutil==2.2
six==1.5.2

